Add sizes and generate set of thumbnails

+++ ImageOptimization/Services/ImageService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
+using System.Text;
 using System.Web.Hosting;
 using ImageOptimization.Models;
 using ImageOptimization.ViewModels;
@@ -15,26 +16,39 @@ namespace ImageOptimization.Services
             if (sourceImage == null)
                 return null;
 
+            // Build sizes component
+            StringBuilder sizes = new StringBuilder();
+
+            foreach(ThumbImage thumb in sourceImage.Thumbnails)
+            {
+                sizes.Append(thumb.RelativePath);
+                sizes.Append(" ");
+                sizes.Append(thumb.Width);
+                sizes.Append("w,");
+            }
 
             return new SourceImageViewModel()
             {
                 ID = sourceImage.ID,
                 FileName = sourceImage.FileName,
                 AltText = sourceImage.AltText,
-                ServerPath = sourceImage.RelativePath,
-                Width = sourceImage.Width,
-                Height = sourceImage.Height,
-                FileFormat = sourceImage.Format
+                FallbackPath = sourceImage.RelativePath,
+                Width = sourceImage.Width.ToString(),
+                Height = sourceImage.Height.ToString(),
+                FileFormat = sourceImage.Format,
+                Thumbnails = sourceImage.Thumbnails,
+                Sizes = sizes.ToString()
             };
         }
 
         /// <summary>
-        /// 
+        /// Generates thumbnail using vips of the source image, always downscales (no upscale), saves the generated
+        /// file to thumbnails folder and returns new ThumbImage Model
         /// </summary>
         /// <param name="src">SourceImage</param>
         /// <param name="width">Width of the image</param>
         /// <param name="height">Height of the image</param>
-        /// <returns></returns>
+        /// <returns>Empty ThumbImage if no file exists or Created ThumbImage</returns>
         internal static ThumbImage GenerateThumbnail(SourceImage src, int width, int? height)
         {
             // Checks, if the file exists
@@ -43,16 +57,20 @@ namespace ImageOptimization.Services
 
             // Create Thumbnail using vips
             Image thumbnail = Image.Thumbnail(src.AbsolutePath, width, height, "down", true);
-            string AppDataFolder = HostingEnvironment.MapPath("~/images");
+            string ThumbnailPath = HostingEnvironment.MapPath("~/thumbnails");
             string FileName = $"th_{thumbnail.Width}x{thumbnail.Height}_{src.FileName}";
-            string path = $"{AppDataFolder}\\{FileName}";
+            string path = $"{ThumbnailPath}\\{FileName}";
 
             // Create corresponding file
             try
             {
-                var file = File.Create(path);
-                file.Close();
-                thumbnail.WriteToFile(path);
+                // If the file already exists, don't create new one
+                if (!File.Exists(path))
+                {
+                    var file = File.Create(path);
+                    file.Close();
+                    thumbnail.WriteToFile(path);
+                }
             }
             catch (Exception e)
             {
@@ -65,7 +83,7 @@ namespace ImageOptimization.Services
                 AbsolutePath = path,
                 AltText = src.AltText,
                 FileName = FileName,
-                RelativePath = $"/images/{FileName}",
+                RelativePath = $"/thumbnails/{FileName}",
                 Height = thumbnail.Height,
                 Width = thumbnail.Width,
                 SourceImageID = src.ID
